ID: 2991614 - Failure mode for Document uploads
In 9.06 and IF the permission are mis-set of the receiving folder, Oscar appears to receive a file to the Documents area, it makes entries in the journa$

My notes:
I can't guarantee this fixes it for all browsers, since there's alot of methods the file upload can happen
depending on the browser..but it's tested in firefox 11, maybe someone try IE/Safari or whatever it is we support.
I think firefox probably is most of our market.