ID: 3534352 - Help link outdated
Changed the links to the OSCAR manual to use a property setting
(resource_base_url) instead of hardcoded links to a page that is
outdated.  Rationale is that each OSCAR version should link to the
appropriate help page - Having OSCAR 10 links redirect to the OSCAR 12
manual page makes as little sense as having OSCAR 12 use the OSCAR 10
manual.

The properties setting would also enable OSPs to link to their own
support material.