ID: 3534352 - Help link outdated - UPDATED
There is an admin feature that allows to update this resource which on
the schedule will override the new property(resource_base_url). But in
the Program Module the Resource link remains what was set in the
property(resource_base_url). Updated so the resource link responds the
same in Program Module as in the schedule.

Change-Id: I83483e13fbbf4b3130772aad0551fc5c5a89eb5b