Fix for HC swipe on new demographic page
The javascript for this page attempts to refer to 'ctx', which is
supposed to contain the page context.  Replaced the reference with
'window.pageContext', but it could also have been done by having the JSP
create the ctx javascript variable.

The other JS scripts in the hcHandler folder are also affected, but not
all cases are fixable with the above method.