oscarEncounter tickler window does not close
Sourceforge ID: 3608171

The tickler window makes a reference to
window.opener.document.forms['encForm']) when saving and closing.
This form is in the main tickler screen, but not in the oscarEncounter
screen. As a result the javascript is interrupted and self.close()
is never reached.

The fix was to put a check in the JS to ensure that the form exists
before triggering the relevant code.